requirements.txt
Flask==1.1.2
SQLAlchemy==1.4.1
Flask-SQLAlchemy==2.5.1
bcrypt==3.2.0
PyJWT==2.0.1
redis==3.5.3
eventlet==0.30.2
Flask-Cors==3.0.10
mysql-connector-python==8.0.23
